+++ unnamed_engine/src/lib.rs
@@ -4,5 +4,6 @@ pub mod renderer;
 pub mod event;
 pub mod gui;
 pub mod input;
+pub mod networking;
 
 pub use log;